Class's moved from ErrorHandling to framework / exception
