+++ src/main/java/com/prolinkli/framework/exception/GlobalExceptionHandler.java
@@ -1,4 +1,5 @@
-package com.prolinkli.core.app.errorhandling;
+package com.prolinkli.framework.exception;
+
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
