+++ src/main/java/com/prolinkli/framework/exception/ResourceNotFoundException.java
@@ -1,4 +1,5 @@
-package com.prolinkli.core.app.errorhandling;
+package com.prolinkli.framework.exception;
+
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
